Modified config to control/switch google analytics with config.yml
Previous behavior was analytics site-wide which is probably the typical
desired behavior. But just in case we want to switch analytics off on a
particular page or post, we can now do so in config.yml with the
`analytics: true|false` switch. Fails to `analytics: false`, but
default setting in config.yml for pages and posts is now `analytics:
true`.